+++ src/components/Interpreter/modules/Meta/Reverse/index.js
@@ -1,0 +1,11 @@
+import * as tf from "@tensorflow/tfjs-core"
+import { of } from "rxjs"
+
+import Symbols from "../../../symbols"
+
+import doc from "./doc.md"
+
+export default {
+    [Symbols.call]: (tensor) => of(tf.reverse(tensor)),
+    [Symbols.doc]: doc
+}